Windows: osilog param size is size_t
The parameter size is size_t which permits pointers to be stored
on 32-bit and 64-bit systems.  Make sure we store it as size_t
instead of truncating to long.

Reviewed-on: http://gerrit.openafs.org/3657
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Tested-by: Jeffrey Altman <jaltman@openafs.org>

+++ src/WINNT/client_osi/osilog.h
@@ -24,7 +24,7 @@ typedef struct osi_logEntry {
 	size_t tid;			/* thread ID */
         unsigned long micros;		/* microsecond-based time stamp */
         char *formatp;			/* format string */
-        long parms[OSI_LOG_MAXPARMS];	/* parms */
+        size_t parms[OSI_LOG_MAXPARMS];	/* parms */
 } osi_logEntry_t;
 
 typedef struct osi_log {
